Support for clients that do not support AirDrop
## Overview

This commit adds the ability to register an arbitrary peer with the services that handle the AirDrop protocol.

When an implementation of the abstract class `AirDropPeer` is registered with the `AirDropService` it becomes discoverable in the AirDrop browser on compatible clients.

An implementation is expected to implement two methods:

 - `CanAcceptFilesAsync` - this method should ask the end user whether they wish to accept files from the sender
 - `OnFileUploadedAsync` - this method should allow the user to download an uploaded file. It is called for each file that was uploaded

Note that because we cannot expect a peer that does not implement AirDrop to support the CPIO file format used by Apple devices the underlying services perform a streaming extraction of the CPIO archive to the intermediate device's file system. `OnFileUploadedAsync` is called for each file present in the archive and an endpoint is provided to allow the target device to download the file.

## CLI implementation

Initial implementation of a peer is provided in the form of a SignalR client and hub, both encapsulated in the `AirDropAnywhere.Cli` project. This is a relatively stock SignalR implementation with two quirks:

 1. It uses bi-directional streaming, built upon `IAsyncEnumerable` and a request/response model is built upon it.
    Request/response is handled by keeping track of a pooled `IValueTaskSource` that is registered when a request is sent and is resolved when the response has been received. A GUID is used as the message identifier and responses are expected to set their `ReplyTo` property to the value of the request's message identifier.
 1. To handle the polymorphic messages sent over the `IAsyncEnumerable` a `PolymorphicJsonConverter` is registered with SignalR that handles the serialization / deserialization of each message by wrapping / unwrapping the payload using a well-known identifier for each message type. These well-known identifiers are registered on the base class using `[PolymorphicJsonInclude]` attribute in much the same way as `ProtoInclude` in protobuf-net

Several messages are implemented to handle the implementation requirements of an `AirDropPeer`:

 - `ConnectMessage` - sent by the client when it initially connects to a server to identify itself
 - `CanAcceptFiles(Request|Response)Message` - sent by the server when a sender tries to send files to the peer
 - `OnFileUploaded(Request|Response)Message` - sent by the server when a file is extracted from the archive sent by a sender to the peer

CLI has been extended to gracefully start/stop either the server or client depending upon the combination of arguments passed when executed. A simple command-line interface has been provided to allow a sender to send files to a peer.

## Other changes

### Infrastructure

In order to support the arbitrary registration of peers `MulticastDnsServer` & `AirDropService` have been slightly adjusted to allow registration and deregistration of peers at runtime. Each peer is tracked in `AirDropService` and is typically registered when a peer connects to the backend and unregistered when it disconnects. `MulticastDnsServer` now uses a TTL of 5 minutes for a peer's mDNS records and multicasts the same records with a TTL of 0s when a peer is unregistered.

`AirDropRouteHandler` has been adjusted so that the host header passed by an AirDrop-compatible device is resolved to the peer that it represents. If the host header does not map to a known peer a 404 is returned. `AirDropRouteHandler` now has the peer passed to it if it was successfully resolved.

### Misc

Rider launch profiles have been added that map to those in `launchSettings.json`