Add support for generating self-signed certificates instead of using the ASP.NET development certificate
This code uses the same mechanisms as that used by the ASP.NET development certificate (see https://github.com/dotnet/aspnetcore/blob/main/src/Shared/CertificateGeneration/CertificateManager.cs), although it _currently_ does not store the certificate it generates.

This is now needed because a self-contained application does not rely upon an installed framework and there is no `dotnet dev-certs` command that can be used to generate the dev certificates!