Tweaks to run on Raspberry Pi / Linux (#15)
After configuring a Raspberry Pi with an RFMON (active monitor) mode capable wifi adapter and running [OWL](https://github.com/seemoo-lab/owl) this PR makes the necessary changes to gracefully run AirDrop Anywhere in that environment.

First, it makes minor tweaks to how listener sockets are bound in the mDNS server (certain socket options are not supported in Linux, only MacOS). Next it tweaks the deployment mechanism to remotely debug on a Raspberry Pi (using a self-contained application so .NET does not need to be installed on the device) and adds support for SSH authentication during that deployment using a YubiKey. Finally, because of the self-contained deployment it adds support for generating a self-signed HTTPS certificate - when the .NET SDK is not installed `dotnet dev-certs` is not available and we are no longer able to rely on the ASP.NET Development Certificate.

Longer term we'll store the HTTPS certificate in the user's certificate store but, for now, we generate a new one each time the application is started.

Good news is that this now works on a Raspberry Pi, partially addressing #11 .

Next up here is to add support for running the application as a systemd daemon.